# Api/Viajes.Api/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Transportes Cuauhtémoc",
    "description" : "Transportes Cuauhtémoc (ASP.NET Core 3.1)",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "security" : [ {
    "Bearer" : [ ]
  } ],
...
  "paths" : {
...
    "/travel-plans" : {
      "get" : {
        "tags" : [ "travel-plans" ],
        "summary" : "Obtiene los planes de viaje del periodo especificado",
        "description" : "Ejecuta el algoritmo de planeación y regresa el resultado correspondiente",
        "operationId" : "GetTravelPlans",
        "parameters" : [ {
          "name" : "startDate",
          "in" : "query",
          "description" : "La primera fecha a considerar",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "endDate",
          "in" : "query",
          "description" : "La última fecha a considerar",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/TravelPlan"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "ProblemDetails" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string",
            "nullable" : true
          },
          "title" : {
            "type" : "string",
            "nullable" : true
          },
          "status" : {
            "type" : "integer",
            "nullable" : true
          },
          "detail" : {
            "type" : "string",
            "format" : "int32",
            "nullable" : true
          },
          "instance" : {
            "type" : "string",
            "nullable" : true
          }
        },
        "additionalProperties" : {
          "type" : "object",
          "additionalProperties" : false
        }
      },
      "BaseEntity" : {
        "required" : [ "active", "created_at", "created_by", "id", "updated_at", "updated_by" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "description" : "El id de esta entidad",
            "format" : "int32"
          },
          "active" : {
            "type" : "boolean",
            "description" : "La entidad se encuentra activa"
          },
          "created_by" : {
            "type" : "integer",
            "description" : "El id del usuario que creo esta entidad",
            "format" : "int32"
          },
          "created_at" : {
            "type" : "string",
            "description" : "La fecha de creación de la entidad",
            "format" : "date-time"
          },
          "updated_by" : {
            "type" : "integer",
            "description" : "El id del último usuario que actualizo esta entidad",
            "format" : "int32"
          },
          "updated_at" : {
            "type" : "string",
            "description" : "La fecha de la última actualización de la entidad",
            "format" : "date-time"
          }
        }
      },
      "NamedEntity" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/BaseEntity"
        }, {
          "required" : [ "description", "name" ],
          "type" : "object",
          "properties" : {
            "name" : {
              "type" : "string",
              "description" : "El nombre de esta entidad"
            },
            "description" : {
              "type" : "string",
              "description" : "El nombre de esta entidad"
            }
          }
        } ]
      },
...
      "TravelPlan" : {
        "type" : "object",
        "properties" : {
          "date" : {
            "type" : "string",
            "description" : "Fecha programada",
            "format" : "date"
          },
          "office_plans" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/OfficePlan"
            }
          }
        },
        "example" : {
          "date" : "2000-01-23",
          "office_plans" : [ {
            "trucks_plan" : {
              "arrivals" : [ null, null ],
              "positioned" : [ "", "" ],
              "completed" : true,
              "demand" : [ {
                "demand" : 6,
                "customer" : 0
              }, {
                "demand" : 6,
                "customer" : 0
              } ]
            },
            "location" : ""
          }, {
            "trucks_plan" : {
              "arrivals" : [ null, null ],
              "positioned" : [ "", "" ],
              "completed" : true,
              "demand" : [ {
                "demand" : 6,
                "customer" : 0
              }, {
                "demand" : 6,
                "customer" : 0
              } ]
            },
            "location" : ""
          } ]
        }
      },
      "OfficePlan" : {
        "type" : "object",
        "properties" : {
          "location" : {
            "$ref" : "#/components/schemas/Location"
          },
          "trucks_plan" : {
            "$ref" : "#/components/schemas/TrucksPlan"
          }
        },
        "example" : {
          "trucks_plan" : {
            "arrivals" : [ null, null ],
            "positioned" : [ "", "" ],
            "completed" : true,
            "demand" : [ {
              "demand" : 6,
              "customer" : 0
            }, {
              "demand" : 6,
              "customer" : 0
            } ]
          },
          "location" : ""
        }
      },
      "TrucksPlan" : {
        "type" : "object",
        "properties" : {
          "demand" : {
            "type" : "array",
            "description" : "Demanda",
            "items" : {
              "$ref" : "#/components/schemas/CustomerDemand"
            }
          },
          "positioned" : {
            "type" : "array",
            "description" : "Posicionados",
            "items" : {
              "$ref" : "#/components/schemas/Truck"
            }
          },
          "arrivals" : {
            "type" : "array",
            "description" : "Llegadas",
            "items" : {
              "$ref" : "#/components/schemas/Truck"
            }
          },
          "completed" : {
            "type" : "boolean",
            "description" : "Completado",
            "nullable" : false
          }
        },
        "example" : {
          "arrivals" : [ null, null ],
          "positioned" : [ "", "" ],
          "completed" : true,
          "demand" : [ {
            "demand" : 6,
            "customer" : 0
          }, {
            "demand" : 6,
            "customer" : 0
          } ]
        }
      },
      "CustomerDemand" : {
        "type" : "object",
        "properties" : {
          "customer" : {
            "type" : "integer",
            "description" : "Cliente",
            "format" : "int32",
            "nullable" : true
          },
          "demand" : {
            "type" : "integer",
            "description" : "Demanda",
            "format" : "int32",
            "nullable" : true
          }
        },
        "example" : {
          "demand" : 6,
          "customer" : 0
        }
      },
      "Truck" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/NamedEntity"
        }, {
          "type" : "object",
          "properties" : {
            "tracking_id" : {
              "type" : "integer",
              "description" : "El id de seguimiento de la unidad",
              "format" : "int32"
            },
            "type" : {
              "type" : "integer",
              "description" : "El id del tipo de unidad",
              "format" : "int32"
            },
            "status" : {
              "type" : "integer",
              "description" : "El estatus del equipo"
            },
            "position" : {
              "$ref" : "#/components/schemas/Position"
            }
          }
        } ]
      },
      "Position" : {
        "required" : [ "coordinates", "timestamp" ],
        "type" : "object",
        "properties" : {
          "timestamp" : {
            "type" : "string",
            "description" : "Timestamp de esta posición",
            "format" : "date-time",
            "nullable" : false
          },
          "coordinates" : {
            "$ref" : "#/components/schemas/Coordinates"
          }
        }
      },
...
      "Coordinates" : {
        "required" : [ "latitude", "longitude" ],
        "type" : "object",
        "properties" : {
          "latitude" : {
            "type" : "number",
            "description" : "Latitud",
            "format" : "double",
            "nullable" : false
          },
          "longitude" : {
            "type" : "number",
            "description" : "Longitud",
            "format" : "double",
            "nullable" : false
          }
        },
        "description" : "Las coordenadas de esta ubicación"
      },
      "Location" : {
        "required" : [ "city", "coordinates", "description", "name" ],
        "description" : "Representa un origen, un destino o un regreso",
        "allOf" : [ {
          "$ref" : "#/components/schemas/NamedEntity"
        }, {
          "type" : "object",
          "properties" : {
            "city" : {
              "type" : "integer",
              "description" : "El id de la ciudad en la que se encuentra esta ubicación",
              "format" : "int32"
            },
            "coordinates" : {
              "$ref" : "#/components/schemas/Coordinates"
            }
          }
        } ]
      },
...
    },
...
    "securitySchemes" : {
      "Bearer" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}